api/api-pynest/pyproject.toml
[project]
name = "api-pynest"
version = "0.1.0"
description = ""
authors = [
    {name = "N-JY",email = "131830774+N-JY@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pynest-api (>=0.3.2,<0.4.0)",
    "starlette (==0.41.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
